Add files via upload
Consider the problem of auctions or penny social. We will refer to it as a simple “Auction.” It is a
conventional approach used for fundraising for a cause. The organizers collect items to be auctioned off
for raising funds. Before the auction, the items for auction are received and arranged each with a bowl
to place the bid. A chairperson is a special person among the organizers. She/he heads the effort and is
the only person who can determine the winner by random drawing at the end of the auction. A set of
bidders buy sheets of tickets with their money. The bidder’s sheet has a stub that identifies the bidder’s
number, and tokens bought.
The bidders examine the items to bid, place the one or more tickets in the bowl in front of the items they
desire to bid for until all the tickets are used. After the auction period ends the chairperson, collects the
bowls, randomly selects a ticket from each item’s bowl to determine the winning bidder for that item. The
item is transferred to the winning bidder. Total money collected is the fund raised by the penny social
auction.